requirements.txt
Flask==2.1.1
fuzzywuzzy==0.18.0
google-api-core==1.31.5
google-auth==1.35.0
google-cloud-vision==2.7.2
googleapis-common-protos==1.56.0
matplotlib-inline==0.1.3
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
pip==22.0.4
scipy==1.8.0
urllib3==1.26.9
Werkzeug==2.1.1
